refactor, add javadoc
